requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
openai==0.27.0
langchain-community==0.0.255
pydantic==1.10.9
python-dotenv==1.0.0
requests==2.28.2